PR Comments
- update eslint to use `.js` files
- use npm instead of yarn
