+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "PhotoSwipe",
   "version": "0.0.1",
-  "private": true,
   "scripts": {
     "start": "node node_modules/react-native/local-cli/cli.js start",
+    "flow": "flow",
     "test": "jest",
-    "lint": "eslint src/"
+    "lint": "eslint . --ext .js"
   },
   "husky": {
     "hooks": {
@@ -14,8 +14,7 @@
   },
   "lint-staged": {
     "*.js": [
-      "eslint",
-      "git add"
+      "eslint"
     ]
   },
   "dependencies": {
